# remove DepthwiseConv type in favor of Conv