Implement @mcabbott's [comment](https://github.com/FluxML/Flux.jl/issues/459#issuecomment-1073238976)

Also adds the assertion 
```julia
  @assert cin % groups == 0 "Input channel dimension must be divisible by groups."
  @assert cout % groups == 0 "Output channel dimension must be divisible by groups."
```
for standard convolutions.

Close #1667, close #459