@mcabbott 2022-04-01
Should what's left move to deprecations.jl too?