@CarloLucibello 2022-04-01
I think it is good to have a convenience constructor like that, helps you understand what a depthwise convolution is