@ToucheSir 2022-04-01
Would you mind running some microbenchmarks on CPU pre- and post-PR? If there's basically no difference between the two, let's merge asap :)