@mcabbott 2022-04-02
Xref https://github.com/FluxML/NNlib.jl/pull/395 perhaps. The present code has 3 layers of multi-threading, not sure how carefully it was ever benchmarked.