@CarloLucibello 2022-04-02
Performance here seems slightly better. We should remove the specialized `depthwiseconv` from NNlib.

## This PR

```julia
julia> using Flux, BenchmarkTools

julia> x = randn(Float32, 128, 128, 32, 32);

julia> dconv = DepthwiseConv((3,3), 32 => 64)
Conv((3, 3), 32 => 64, groups=32)  # 640 parameters

julia> Threads.nthreads()
1

julia> @btime $dconv($x);
  169.259 ms (320 allocations: 265.55 MiB)
```

```julia
julia> Threads.nthreads()
4

julia> @btime $dconv($x);
  96.917 ms (366 allocations: 317.88 MiB)
```

## On Master

```julia
julia> using Flux, BenchmarkTools

julia> x = randn(Float32, 128, 128, 32, 32);

julia> dconv = DepthwiseConv((3,3), 32 => 64)
DepthwiseConv((3, 3), 32 => 64)  # 640 parameters

julia> Threads.nthreads()
1

julia> @btime $dconv($x);
  180.530 ms (22 allocations: 265.51 MiB)
```

```julia
julia> Threads.nthreads()
4

julia> @btime $dconv($x);
  101.642 ms (37 allocations: 317.83 MiB)
```